rename remaining utility namespaces
