+++ composer.json
@@ -33,30 +33,30 @@
             "src/Utils/HSV/utils.php",
             "src/Utils/HWB/conversions.php",
             "src/Utils/HWB/utils.php",
-            "src/Utils/LAB/conversions.php",
-            "src/Utils/LAB/utils.php",
-            "src/Utils/LCH/conversions.php",
-            "src/Utils/LCH/utils.php",
+            "src/Utils/Lab/conversions.php",
+            "src/Utils/Lab/utils.php",
+            "src/Utils/Lch/conversions.php",
+            "src/Utils/Lch/utils.php",
             "src/Utils/LinP3/conversions.php",
             "src/Utils/LinP3/utils.php",
             "src/Utils/LinProPhoto/conversions.php",
             "src/Utils/LinProPhoto/utils.php",
             "src/Utils/LinRGB/conversions.php",
             "src/Utils/LinRGB/utils.php",
-            "src/Utils/OkLAB/conversions.php",
-            "src/Utils/OkLAB/utils.php",
-            "src/Utils/OkLCH/conversions.php",
-            "src/Utils/OkLCH/utils.php",
+            "src/Utils/OkLab/conversions.php",
+            "src/Utils/OkLab/utils.php",
+            "src/Utils/OkLch/conversions.php",
+            "src/Utils/OkLch/utils.php",
             "src/Utils/P3/conversions.php",
             "src/Utils/P3/utils.php",
             "src/Utils/ProPhoto/conversions.php",
             "src/Utils/ProPhoto/utils.php",
             "src/Utils/RGB/conversions.php",
             "src/Utils/RGB/utils.php",
-            "src/Utils/XYZD50/conversions.php",
-            "src/Utils/XYZD50/utils.php",
-            "src/Utils/XYZD65/conversions.php",
-            "src/Utils/XYZD65/utils.php"
+            "src/Utils/XyzD50/conversions.php",
+            "src/Utils/XyzD50/utils.php",
+            "src/Utils/XyzD65/conversions.php",
+            "src/Utils/XyzD65/utils.php"
         ],
         "psr-4": {
             "Atomicptr\\Color\\": "src/"
